refactor: break down events json_log.rs into focused modules
- Split 480-line json_log.rs into organized module hierarchy
- All files under 200 lines (largest: 156 lines)
- Group by function: config, writer, rotation, subscriber
- Maintain all functionality and API compatibility
